`ResourceMap` -> `ResourceMap[int]`.
Fixes pytype.

+++ axlearn/cloud/common/quota.py
@@ -19,9 +19,9 @@ class QuotaInfo:
 
     # A mapping from resource type to total resource limits.
     total_resources: ResourceMap[int]
-    # A nested mapping. Key is project identifier, value is mapping from resource type to allocated
-    # amount of resources.
-    project_resources: ProjectResourceMap[int]
+    # A nested mapping. Key is project identifier, value is mapping from resource type to
+    # per-project resource proportions.
+    project_resources: ProjectResourceMap[float]
 
 
 class QuotaFn(Protocol):
